test: inline wall clock rate limiter tests
